Delete post with observation!
After confirming the deletion of the post in the modal window I could not find an easy way to update the state of the Home component. I tried to use the useEffect hook, but it did not work. I decided to use the window.location.reload() method to reload the page after deleting the post. I know that this is not the best solution, but do to time constrains I decided to use it. I will refactor it in the future after feedback from the reviewer.